requirements.txt
google-generativeai==0.3.2
python-dotenv==1.0.0
streamlit==1.31.1
Pillow==10.2.0
